♻️ refactor: remove fully qualified class names e adiciona imports corretos
- Remove referências FQN como \Illuminate\, \App\Models\, \Exception
- Adiciona declarações use apropriadas (Exception, InvalidArgumentException, Override, etc)
- Melhora legibilidade do código seguindo PSR standards
- Atualiza annotations PHPDoc com tipos simplificados
- Remove redundâncias de casting explícito ((int), (float))
- Padroniza implementação de interfaces (Stringable)

Arquivos afetados:
- [RedisCacheInvalidator.php](http://_vscodecontentref_/17)
- [ArticleRepositoryInterface.php](http://_vscodecontentref_/18)
- [CreateArticleDTO.php](http://_vscodecontentref_/19)
- [ArticleRankingController.php](http://_vscodecontentref_/20)
- [TrackArticleView.php](http://_vscodecontentref_/21)
- app/Http/Requests/*.php (4 arquivos)
- [Article.php](http://_vscodecontentref_/22)
- app/Providers/*.php (2 arquivos)
- [ArticleRepository.php](http://_vscodecontentref_/23)
- app/Services/*.php (2 arquivos)
- [Url.php](http://_vscodecontentref_/24)
- routes/web.php

+++ app/Services/ArticleRankingService.php
@@ -101,7 +101,7 @@ class ArticleRankingService implements ArticleRankingServiceInterface
 
         /** @var array<string, float> $topArticle */
         $topArticle = Redis::zrevrange(self::RANKING_KEY, 0, 0, ['withscores' => true]);
-        $topScore = empty($topArticle) ? 0.0 : (float) array_values($topArticle)[0];
+        $topScore = empty($topArticle) ? 0.0 : array_values($topArticle)[0];
 
         /** @var array<string, float> $allScores */
         $allScores = Redis::zrange(self::RANKING_KEY, 0, -1, ['withscores' => true]);
